changed size of bot, fixed hitbox issues

+++ cs-base-station-interface/src/main/java/basestation/vision/VisionObject.java
@@ -37,11 +37,11 @@ public class VisionObject {
     }
 
     public VisionObject(VisionSystem vs, int vid, VisionCoordinate myCoord,
-                        int size) {
+                        float size) {
         this.vs = vs;
         this.vid = vid;
         this.coord = myCoord;
-        this.size = (float) size;
+        this.size = size;
     }
 
     public VisionObject(VisionObject old, VisionCoordinate newCoord) {
